(#236) refactor: DetailIssueCell 파일 분리
